pyproject.toml
[tool.poetry]
name = "community-detection-complex-networks"
version = "0.1.0"
description = "Analysis of community structures in complex networks using SBM with Infomap, Louvain, and Leiden algorithms."
authors = ["Simone Gigante <simonegigante17@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
networkx = "^3.3"
python-igraph = "^0.11.4"
infomap = "^2.7.1"
scikit-learn = "^1.4.2"
matplotlib = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

